fix y axis offset now that safeArea is not centred
