+++ src/core/layout/scrollable-list/scrollable-list-handlers.js
@@ -6,7 +6,11 @@
  */
 /* eslint-disable no-console */
 
-/* gelButton's clickable zone extends beyond rexUI's mask [CGPROD-2795]*/
+import { getMetrics } from "../../scaler.js";
+import fp from "../../../../lib/lodash/fp/fp.js";
+
+const GRID_NAME = "grid";
+
 const handleIfVisible = (gelButton, scene) => {
     const outerContainer = gelButton.rexContainer.parent.getTopmostSizer();
     const mouseY = scene.input.y;
@@ -17,8 +21,70 @@ const handleIfVisible = (gelButton, scene) => {
     if (mouseY >= topY && mouseY <= bottomY) onClickPlaceholder(gelButton);
 };
 
-const assetKey = (key, assetKeys) => [assetKeys.prefix, key].join(".");
-
 const onClickPlaceholder = gelButton => console.log(`Clicked ${gelButton.config.id}`);
 
-export { handleIfVisible, assetKey };
+const updatePanelOnScroll = panel => () => {
+    if (!panel.a11yWrapper) return;
+
+    if (!panel.a11yWrapper.style.cssText) {
+        panel.a11yWrapper.style.position = "absolute";
+        panel.a11yWrapper.style.top = "0px";
+    }
+    const space = panel.space.top;
+    const totalItemsHeight = getItemsHeight(panel);
+    const panelInnerHeight = panel.height - space;
+    const yOffset = -(totalItemsHeight - panelInnerHeight) * panel.t * getMetrics().scale;
+    panel.a11yWrapper.style.top = yOffset + "px";
+};
+
+const getItemsHeight = panel => {
+    const items = getPanelItems(panel);
+    return items.length * (items[0].height + panel.space.top);
+};
+
+const getPanelItems = panel => panel.getByName(GRID_NAME, true).getElement("items");
+
+const updatePanelOnFocus = panel => rexLabel => {
+    const visibleBounds = getVisibleRangeBounds(panel);
+    const itemBounds = getItemBounds(panel, rexLabel);
+    fp.cond([
+        [
+            (visible, item) => item.lower < visible.lower,
+            (visible, item) => updatePanelOffset(panel, item.lower - visible.lower),
+        ],
+        [
+            (visible, item) => item.upper > visible.upper,
+            (visible, item) => updatePanelOffset(panel, item.upper - visible.upper),
+        ],
+        [() => true, () => {}],
+    ])(visibleBounds, itemBounds);
+};
+
+const getVisibleRangeBounds = panel => {
+    const itemsHeight = getItemsHeight(panel);
+    const maxOffset = getMaxOffset(panel);
+    const lower = maxOffset * panel.t;
+    const upper = itemsHeight - (maxOffset - lower);
+    return { lower, upper };
+};
+
+const getItemBounds = (panel, rexLabel) => {
+    const idx = getPanelItems(panel).findIndex(item => item === rexLabel);
+    const lower = idx * rexLabel.height + idx * panel.space.top;
+    const upper = lower + rexLabel.height;
+    return { lower, upper };
+};
+
+const updatePanelOffset = (panel, offset) => {
+    const maxOffset = getMaxOffset(panel);
+    const tDelta = offset / maxOffset;
+    panel.t += tDelta;
+};
+
+const getMaxOffset = panel => {
+    const visibleWindowHeight = panel.minHeight - panel.space.top * 2;
+    const itemsHeight = getItemsHeight(panel);
+    return itemsHeight - visibleWindowHeight;
+};
+
+export { handleIfVisible, updatePanelOnFocus, updatePanelOnScroll };
